requirements.txt
torch
torchvision
pathlib 
matplotlib
numpy
argparse
tqdm
pillow
gdown